server/fastapi/pyproject.toml
[project]
name = "zelfhosted-api"
version = "0.1.0"
description = "AI-powered backend for Zelfhosted"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "uvicorn[standard]>=0.38.0",
    "langgraph>=0.4.0",
    "langchain-openai>=0.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "requests>=2.32.0",
    "requests-oauthlib>=1.3.0",
    "feedparser>=6.0.0",
    "nyct-gtfs>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
